requirements.txt
annotated-types>=0.7,<1
beautifulsoup4>=4.13,<5
cachetools>=5.5,<6
certifi>=2025.8,<2026
cloudscraper>=1.2,<2
charset-normalizer>=3.4,<4
colorama>=0.4,<1

google-api-python-client>=2.179,<3
google-auth-oauthlib>=1.2,<2
google-genai>=1.66,<2

httplib2>=0.22,<1
idna>=3.10,<4
iniconfig>=2.1,<3
oauthlib>=3.3,<4
packaging>=25,<26
pluggy>=1.6,<2
pydantic>=2.12,<3
Pygments>=2.19,<3
pyparsing>=3.2,<4
pytest>=8.4,<9
python-dateutil>=2.9,<3
python-dotenv>=1.1,<2
requests>=2.32,<3
requests-oauthlib>=2.0,<3
soupsieve>=2.7,<3
tqdm>=4.67,<5
typing-inspection>=0.4,<1
typing_extensions>=4.14,<5
uritemplate>=4.2,<5
urllib3>=2.5,<3
